@@ -1,7 +1,7 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFrameworks>netstandard2.1</TargetFrameworks>
+		<TargetFrameworks>net8.0;netstandard2.1</TargetFrameworks>
 		<Company>Digital Garage</Company>
 		<Version>4.3.9</Version>
 		<Copyright>Copyright © Digital Garage 2017</Copyright>
@@ -13,10 +13,10 @@
 		<RepositoryUrl>https://github.com/dgarage/NBXplorer</RepositoryUrl>
 		<RepositoryType>git</RepositoryType>
 		<PackageReadmeFile>README.md</PackageReadmeFile>
-		<LangVersion>10.0</LangVersion>
+		<LangVersion>12</LangVersion>
 	</PropertyGroup>
-  <PropertyGroup Condition=" '$(TargetFramework)' == 'netstandard2.0' ">
-    <DefineConstants>$(DefineConstants);NO_SPAN</DefineConstants>
+  <PropertyGroup Condition=" '$(TargetFramework)' == 'netstandard2.1' ">
+    <DefineConstants>$(DefineConstants);NO_RECORD</DefineConstants>
   </PropertyGroup>
         <PropertyGroup Condition=" '$(Configuration)' == 'Release' ">
 		  <PublishRepositoryUrl>true</PublishRepositoryUrl>
@@ -27,7 +27,7 @@
 		  <NoWarn>$(NoWarn);1591;1573;1572;1584;1570;3021</NoWarn>
         </PropertyGroup>
 	<ItemGroup>
-		<PackageReference Include="NBitcoin" Version="8.0.13" />
+		<PackageReference Include="NBitcoin" Version="8.0.17" />
 		<PackageReference Include="NBitcoin.Altcoins" Version="4.0.8" />
 	</ItemGroup>
   <ItemGroup>
