@@ -4,10 +4,10 @@ using NBitcoin;
 
 namespace NBXplorer.DerivationStrategy
 {
-	public class P2SHDerivationStrategy : DerivationStrategyBase
+	public class P2SHDerivationStrategy : StandardDerivationStrategyBase
 	{
 		bool addSuffix;
-		internal P2SHDerivationStrategy(DerivationStrategyBase inner, bool addSuffix):base(inner.AdditionalOptions)
+		internal P2SHDerivationStrategy(StandardDerivationStrategyBase inner, bool addSuffix):base(inner.AdditionalOptions)
 		{
 			if(inner == null)
 				throw new ArgumentNullException(nameof(inner));
@@ -15,7 +15,7 @@ namespace NBXplorer.DerivationStrategy
 			this.addSuffix = addSuffix;
 		}
 
-		public DerivationStrategyBase Inner
+		public StandardDerivationStrategyBase Inner
 		{
 			get; set;
 		}
@@ -30,24 +30,17 @@ namespace NBXplorer.DerivationStrategy
 			}
 		}
 
-		public override Derivation GetDerivation()
-		{
-			var derivation = Inner.GetDerivation();
-			return new Derivation()
-			{
-				ScriptPubKey = derivation.ScriptPubKey.Hash.ScriptPubKey,
-				Redeem = derivation.Redeem ?? derivation.ScriptPubKey
-			};
-		}
-
 		public override IEnumerable<ExtPubKey> GetExtPubKeys()
 		{
 			return Inner.GetExtPubKeys();
 		}
 
-		public override DerivationStrategyBase GetChild(KeyPath keyPath)
+		public override Derivation GetDerivation(KeyPath keyPath)
 		{
-			return new P2SHDerivationStrategy(Inner.GetChild(keyPath), addSuffix);
+			var derivation = Inner.GetDerivation(keyPath);
+			return new Derivation(
+				 derivation.ScriptPubKey.Hash.ScriptPubKey,
+				 derivation.Redeem ?? derivation.ScriptPubKey);
 		}
 	}
 }
