@@ -158,22 +158,27 @@ namespace NBXplorer.Backend
 
 		internal record ScriptInsert(string code, string wallet_id, string script, string addr, bool used);
 		internal record DescriptorScriptInsert(string descriptor, int idx, string script, string metadata, string addr, bool used);
-		public async Task<int> GenerateAddresses(DerivationStrategyBase strategy, DerivationFeature derivationFeature, GenerateAddressQuery query = null)
+
+		public Task<int> GenerateAddresses(DerivationStrategyBase strategy, DerivationFeature derivationFeature, GenerateAddressQuery query = null)
+		=> GenerateAddresses(strategy, strategy.GetLineFor(KeyPathTemplates, derivationFeature), query);
+		public async Task<int> GenerateAddresses(DerivationStrategyBase strategy, DerivationLine derivationLine, GenerateAddressQuery query = null)
 		{
 			await using var connection = await connectionFactory.CreateConnection();
-			return await GenerateAddressesCore(connection, strategy, derivationFeature, query);
+			return await GenerateAddressesCore(connection, strategy, derivationLine, query);
 		}
 
 		record GapNextIndex(long gap, long next_idx);
-		internal async Task<int> GenerateAddressesCore(DbConnection connection, DerivationStrategyBase strategy, DerivationFeature derivationFeature, GenerateAddressQuery query)
+		public Task<int> GenerateAddressesCore(DbConnection connection, DerivationStrategyBase strategy, DerivationFeature derivationFeature, GenerateAddressQuery query)
+=> GenerateAddressesCore(connection, strategy, strategy.GetLineFor(KeyPathTemplates, derivationFeature), query);
+		internal async Task<int> GenerateAddressesCore(DbConnection connection, DerivationStrategyBase strategy, DerivationLine derivationLine, GenerateAddressQuery query)
 		{
-			var descriptorKey = GetDescriptorKey(strategy, derivationFeature);
+			var descriptorKey = GetDescriptorKey(strategy, derivationLine.Feature);
 			var walletKey = GetWalletKey(strategy, Network);
 			var gapNextIndex = await GetGapAndNextIdx(connection, descriptorKey);
 			long toGenerate = ToGenerateCount(query, gapNextIndex?.gap);
 			if (gapNextIndex is not null && toGenerate == 0)
 				return 0;
-			var keyTemplate = KeyPathTemplates.GetKeyPathTemplate(derivationFeature);
+			var keyTemplate = (derivationLine as KeyPathTemplateDerivationLine)?.KeyPathTemplate;
 			if (gapNextIndex is null)
 			{
 				// Let's generate the wallet
@@ -187,7 +192,7 @@ namespace NBXplorer.Backend
 						metadata = Serializer.ToString(new LegacyDescriptorMetadata()
 						{
 							Derivation = strategy,
-							Feature = derivationFeature,
+							Feature = derivationLine.Feature,
 							KeyPathTemplate = keyTemplate,
 							Type = LegacyDescriptorMetadata.TypeName
 						}),
@@ -203,16 +208,15 @@ namespace NBXplorer.Backend
 			do
 			{
 				var nextIndex = gapNextIndex.next_idx;
-				var line = strategy.GetLineFor(keyTemplate);
 				var scriptpubkeys = new Script[toGenerate];
 				var linesScriptpubkeys = new DescriptorScriptInsert[toGenerate];
 				Parallel.For(nextIndex, nextIndex + toGenerate, i =>
 				{
-					var derivation = line.Derive((uint)i);
+					var derivation = derivationLine.Derive((uint)i);
 					scriptpubkeys[i - nextIndex] = derivation.ScriptPubKey;
 
 					var addr = derivation.ScriptPubKey.GetDestinationAddress(Network.NBitcoinNetwork);
-					JObject metadata = GetDescriptorScriptMetadata(strategy, line.KeyPathTemplate.GetKeyPath((int)i, false), derivation, addr);
+					JObject metadata = GetDescriptorScriptMetadata(strategy, derivation, keyTemplate?.GetKeyPath((uint)i), addr);
 					linesScriptpubkeys[i - nextIndex] = new DescriptorScriptInsert(
 						descriptorKey.descriptor,
 						(int)i,
@@ -303,7 +307,7 @@ namespace NBXplorer.Backend
 								});
 		}
 
-		private JObject GetDescriptorScriptMetadata(DerivationStrategyBase strategy, KeyPath keyPath, Derivation derivation, BitcoinAddress addr)
+		private JObject GetDescriptorScriptMetadata(DerivationStrategyBase strategy, Derivation derivation, KeyPath keyPath, BitcoinAddress addr)
 		{
 			JObject metadata = null;
 			if (derivation.Redeem?.ToHex() is string r)
@@ -316,7 +320,7 @@ namespace NBXplorer.Backend
 			{
 				if (!strategy.Unblinded())
 				{
-					var blindingKey = NBXplorerNetworkProvider.LiquidNBXplorerNetwork.GenerateBlindingKey(strategy, keyPath, addr.ScriptPubKey, Network.NBitcoinNetwork);
+					var blindingKey = NBXplorerNetworkProvider.LiquidNBXplorerNetwork.GenerateBlindingKey(strategy, keyPath, derivation.ScriptPubKey, Network.NBitcoinNetwork);
 					var blindedAddress = new BitcoinBlindedAddress(blindingKey.PubKey, addr);
 					metadata ??= new JObject();
 					metadata.Add(new JProperty("blindedAddress", blindedAddress.ToString()));
@@ -458,9 +462,6 @@ namespace NBXplorer.Backend
 				JOIN wallets w USING(wallet_id)", parameters);
 			foreach (var r in rows)
 			{
-				// This might be the case for a derivation added by a different indexer
-				if (r.derivation is not null && r.keypath is null)
-					continue;
 				BitcoinAddress addr = GetAddress(r);
 				bool isExplicit = r.derivation is null;
 				bool isDescriptor = !isExplicit;
@@ -491,11 +492,10 @@ namespace NBXplorer.Backend
 				ki.ScriptPubKey = script;
 				ki.TrackedSource = trackedSource;
 				ki.Feature = DerivationFeature.Deposit;
-				if (keypath is not null)
-				{
+				if (r.feature is not null)
 					ki.Feature = Enum.Parse<DerivationFeature>(r.feature, true);
+				if (r.idx is not null)
 					ki.Index = (int)r.idx;
-				}
 				ki.Redeem = redeem is null ? null : Script.FromHex(redeem);
 				result.Add(script, ki);
 			}
@@ -996,11 +996,10 @@ namespace NBXplorer.Backend
 				if (ki.TrackedSource is DerivationSchemeTrackedSource a)
 				{
 					descriptorKey = GetDescriptorKey(a.DerivationStrategy, ki.Feature);
-					var derivation = a.DerivationStrategy.GetDerivation(ki.KeyPath);
 					metadata = GetDescriptorScriptMetadata(
 								a.DerivationStrategy,
+								new Derivation(ki.ScriptPubKey, ki.Redeem),
 								ki.KeyPath,
-								derivation,
 								addr);
 				}
 
